reorganize common
