+++ environment.yml
@@ -2,7 +2,7 @@ name: rubicon-ml-dev
 channels:
   - conda-forge
 dependencies:
-  - python>=3.8
+  - python>=3.7
   - pip
 
   - click>=7.1
